cleanup types

+++ src/snake.cpp
@@ -37,8 +37,8 @@ void Snake::UpdateHead() {
   }
 
   // Wrap the Snake around to the beginning if going off of the screen.
-  head_x = fmod(head_x + grid_width, grid_width);
-  head_y = fmod(head_y + grid_height, grid_height);
+  head_x = (float) fmod(head_x + (float) grid_width, grid_width);
+  head_y = (float) fmod(head_y + (float) grid_height, grid_height);
 }
 
 void Snake::UpdateBody(SDL_Point &current_head_cell, SDL_Point &prev_head_cell) {
